# perf(modeling): cache poly3.measureBoundingSphere for BSP performance