A number of performance issues were pointed out by @briansturgill. One of them was caching of `measureBoundingSphere` in the BSP boolean operations (#1094). BSP operations use measureBoundingSphere to determine if objects overlap.

By adding caching to `poly3.measureBoundingSphere`, we can get a significant speed up in BSP operations. This uses the same caching code as in the `measurements` package, but applied to the `poly3` package.

For unioning two overlapping spheres with 100 segments, we get **49,366** cache misses (measurements), and **97,030,301** cache hits saved! That's a 99.95% cache hit rate. :grimacing: 

This results in major performance improvements as measured in BSP union operations. Comparing the last release 2.5.10, master (with no-hypot #1099), and measureBoundingSphere caching:

|  | 2.5.10 | master | caching |
| --- | --- | --- | --- |
| disjoint (x=4) | 480ms | 480ms | 480ms |
| touching (x=2) | 8500ms | 5800ms | 3800ms |
| overlap (x=1) | 8900ms | 5850ms | 4000ms |
| same (x=0) | 7300ms | 4800ms | 3700ms |
| inside (x=0) | 4850ms | 3300ms | 2500ms |

So roughly speaking, removing `hypot` gave a ~30% improvement, and caching gives ANOTHER ~30% performance improvement. This gives a total speedup in BSP operations of 48-55% this month! :sunglasses: 


### All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../pulls) for the same update/change?
* [x] Does your submission pass tests?