@z3dev 2022-06-10
@platypii nice

I have some changes on branch perf-booleans which speed up the calculation of bounding sphere. You could merge those into these changes and improve even further.