@platypii 2022-06-10
@z3dev that's interesting, I'm looking at that branch now, and it actually removes a call to `poly3.measureBoundingSphere` so I will have to test to see how these two branches interact with each other on performance.

Another thought I had: why are we using bounding spheres and not bounding boxes? At least in the case of poly3, the bounding sphere is calculated as the sphere containing the bounding box. So the sphere will always be larger than the box. If that's the case, it would be more precise if we checked the smaller bounding box instead of the sphere for checking whether two poly3's have overlap. I'll do some tests.