@briansturgill 2022-06-10
> This results in major performance improvements as measured in BSP union operations. Comparing the last release 2.5.10, master (with no-hypot #1099), and measureBoundingSphere caching:
> 	2.5.10 	master 	caching
> disjoint (x=4) 	480ms 	480ms 	480ms
> touching (x=2) 	8500ms 	5800ms 	3800ms
> overlap (x=1) 	8900ms 	5850ms 	4000ms
> same (x=0) 	7300ms 	4800ms 	3700ms
> inside (x=0) 	4850ms 	3300ms 	2500ms

I'd like to get access to that test program if I could (to translate and run against CSharpCAD).
Brian