@platypii 2022-06-10
@briansturgill thanks! Here is the test script I used. It's a simple union of two spheres at different x distances. I also include a smaller sphere inside a larger sphere test since it exhibits slightly different performance. I'm measuring specifically the `union` call.

<details>
<summary>spheres.js</summary>

```js
const jscad = require('@jscad/modeling')
const { sphere } = jscad.primitives
const { union } = jscad.booleans

const segments = 100

const main = () => {
  test(4, 1, "disjoint")
  test(2, 1, "touching")
  test(1, 1, "overlap")
  test(0, 1, "same")
  test(0, 0.5, "inside")
  return sphere()
}

const test = (x, radius, name) => {
  const s1 = sphere({ segments })
  const s2 = sphere({ segments, center: [x, 0, 0], radius })
  const startTime = performance.now()
  const out = union(s1, s2)
  console.log(`${name} ${performance.now() - startTime} ms`)
  return out
}

module.exports = { main }
```

</details>