@platypii 2022-06-11
@z3dev nice one on `perf-booleans` branch! The fastest BSP I've seen is a cached version of your `perf-booleans` branch. It required some tweaking, and loses some of your memory savings, but this gives the best performance I've seen yet!

|  | 2.5.10 | master | poly3-cache | perf-booleans | perf-booleans-cache |
| --- | --- | --- | --- | --- | --- |
| disjoint (x=4) | 480ms | 480ms | 480ms | 480ms | 480ms |
| touching (x=2) | 8500ms | 5800ms | 3800ms | 4530ms | 3570ms |
| overlap (x=1) | 8900ms | 5850ms | 4000ms | 4820ms | 3840ms |
| same (x=0) | 7300ms | 4800ms | 3700ms | 3980ms | 3160ms |
| inside (x=0) | 4850ms | 3300ms | 2500ms | 2580ms | 2170ms |

This would push the total improvement this month to 55-58% faster boolean operations :exploding_head: