@z3dev 2022-06-11
> Another thought I had: why are we using bounding spheres and not bounding boxes? At least in the case of poly3, the bounding sphere is calculated as the sphere containing the bounding box. So the sphere will always be larger than the box. If that's the case, it would be more precise if we checked the smaller bounding box instead of the sphere for checking whether two poly3's have overlap. I'll do some tests.

Simply put... the polygon is being tested to see if it can be cut by the plane. It uses the distance away from 0,0,0 to see if there's an overlap. The sphere is being calculated to get the **largest** possible delta (radius).