@briansturgill 2022-06-11
@platypii Thanks for the script, timing was wonderful as I was just getting around to doing the optimizing.
Note, I extended the script to do Subtract and Intersect too.

Results of optimization of by simplifying hypot, inlining critical checking and caching Geom2/3 BoundingBox and Poly3 BoundingSphere (CSharpCAD doesn't have measureBoundingBox, but rather puts it on the Geom2/Geom3 objects.)

![image](https://user-images.githubusercontent.com/7350143/173203822-f74cdcc3-7669-4021-81bd-ca214811d303.png)