@platypii 2022-06-11
@z3dev I copied your memory-optimized measureBoundingSphere from #1107 back here into `poly3.measureBoundingSphere`. 

This is the fastest BSP, but does include a breaking change to the API. The performance benefit of changing the api to return `vec4` instead of `[vec3, number]` is:

|  | [vec3, number] | vec4 |
| --- | --- | --- |
| disjoint (x=4) | 474ms | 470ms |
| touching (x=2) | 3950ms | 3650ms |
| overlap (x=1) | 4000ms | 3800ms |
| same (x=0) | 3460ms | 3330ms |
| inside (x=0) | 2290ms | 2140ms |

Which means we get 4-8% improvement in performance from changing the measureBoundingSphere API. Do I think anyone is using `poly3.measureBoundingSphere`? No, almost certainly not. So I think we should change it, personally. But I am open to suggestions here @z3dev.